move scripts out of algolia package
Was getting:

Module not found: Your application tried to access fs. While this module is usually interpreted as a Node builtin, your resolver is running inside a non-Node resolution context where such builtins are ignored. Since fs isn't otherwise declared in your dependencies, this makes the require call ambiguous and unsound.`
