+++ apps/asap-cli/src/scripts/algolia/get-settings.ts
@@ -1,6 +1,7 @@
 /* istanbul ignore file */
 import algoliasearch from 'algoliasearch';
 import fs from 'fs/promises';
+import { resolve } from 'path';
 
 export type GetSettings = {
   algoliaAppId: string;
@@ -13,7 +14,7 @@ export const getAlgoliaSettings = async ({
   algoliaCiApiKey,
   indexName,
 }: GetSettings): Promise<void> => {
-  const path = 'packages/algolia';
+  const path = resolve(__dirname, '../../../../../schema');
   const client = algoliasearch(algoliaAppId, algoliaCiApiKey);
   const index = client.initIndex(indexName);
   const { replicas: _, ...indexSettings } = await index.getSettings();
