+++ apps/asap-cli/src/scripts/algolia/index.ts
