+++ apps/asap-cli/src/scripts/algolia/move-index.ts
