+++ apps/asap-cli/src/scripts/algolia/remove-index.ts
