+++ apps/asap-cli/src/scripts/algolia/remove-records.ts
