+++ packages/algolia/src/filters.ts
@@ -30,5 +30,6 @@ export const getEventFilters = (
     ...(filters.before ? [timeFilter(filters.before, '<')] : []),
     ...(filters.after ? [timeFilter(filters.after, '>')] : []),
   ];
+
   return getFilter(filterList, constraints);
 };
