+++ schema/algolia-schema.json
@@ -44,7 +44,9 @@
   "attributesToRetrieve": null,
   "unretrievableAttributes": null,
   "optionalWords": null,
-  "queryLanguages": ["en"],
+  "queryLanguages": [
+    "en"
+  ],
   "attributesForFaceting": [
     "teams.id",
     "documentType",
@@ -82,6 +84,11 @@
   "highlightPreTag": "<em>",
   "highlightPostTag": "</em>",
   "snippetEllipsisText": "",
-  "alternativesAsExact": ["ignorePlurals", "singleWordSynonym"],
-  "indexLanguages": ["en"]
+  "alternativesAsExact": [
+    "ignorePlurals",
+    "singleWordSynonym"
+  ],
+  "indexLanguages": [
+    "en"
+  ]
 }
